requirements.txt
aiohttp==3.9.0
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
aredis==1.1.8
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
DateTime==5.2
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.2
fastapi-cache2==0.2.1
freezegun==1.2.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
orjson==3.9.9
packaging==23.2
pendulum==2.1.2
pluggy==1.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.3
zope.interface==6.1
